requirements.txt
asgiref==3.8.1
astroid==3.3.5
certifi==2024.8.30
charset-normalizer==3.4.0
coreapi==2.3.3
coreschema==0.0.4
dill==0.3.9
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
idna==3.10
isort==5.13.2
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mccabe==0.7.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
pylint==3.3.1
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
requests==2.32.3
sqlparse==0.5.1
tomli==2.0.2
tomlkit==0.13.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
